correcciones consulta_productos y generacion de pdf

+++ store-invoice-api/src/main/java/com/storeinvoice/storeinvoiceapi/application/service/liquidacion/mapper/ProductoMapper.java
@@ -2,7 +2,6 @@ package com.storeinvoice.storeinvoiceapi.application.service.liquidacion.mapper;
 
 import com.storeinvoice.storeinvoiceapi.application.dto.ProductoPedidoDTO;
 import com.storeinvoice.storeinvoiceapi.domain.model.Producto;
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -31,8 +30,8 @@ public final class ProductoMapper {
                 producto.idProducto(),
                 producto.nombre(),
                 producto.cantidad(),
-                BigDecimal.valueOf(producto.precioUnitario()),
-                BigDecimal.valueOf(producto.subtotal())
+                producto.precioUnitario(),
+                producto.subtotal()
         );
     }
 }
